Rivals greater than skin ID 256 can now properly be added in GP

Closes ring-racers#92

See merge request kart-krew-dev/ring-racers-internal!2867

+++ src/k_grandprix.cpp
@@ -316,7 +316,7 @@ void K_InitGrandPrixBots(void)
 				// Intentionally referenced before (currently dummied out) unlock check. Such a tease!
 				if (rivalnum != -1 && grabskins[(UINT16)rivalnum] != MAXSKINS)
 				{
-					botskinlist[botskinlistpos++] = (UINT8)rivalnum;
+					botskinlist[botskinlistpos++] = (UINT16)rivalnum;
 					grabskins[(UINT16)rivalnum] = MAXSKINS;
 				}
 			}
